Add travis config
