+++ .travis.yml
@@ -1,0 +1,11 @@
+sudo: false
+language: node_js
+node_js:
+  - "8"
+install:
+  - npm install
+script:
+  - npm test:unit --silent
+cache:
+  directories:
+    - node_modules
