+++ README.md
@@ -1,4 +1,22 @@
-# manager
+# User Manager 🤩
+
+Created a reusable widget for managing users.
+* Manager is able to add users (last name, first name, username)
+* Manager is able to see the list of users
+* Manager is able to change first name and last name of any user
+* Manager is not able to change the username
+* Manager is able to filter user list using the search input
+* Username should be between 4-5 symbols and unique
+* First and last name shouldn’t be empty
+* Widget is reusable and can be used in different places
+
+## Demo
+https://user-manager-db.firebaseapp.com/
+
+## Stack of technology
+* Vue
+* Firebase
+* Jest
 
 ## Project setup
 ```
@@ -20,5 +38,7 @@ yarn build
 yarn lint
 ```
 
-### Customize configuration
-See [Configuration Reference](https://cli.vuejs.org/config/).
+### Runs unit tests
+```
+yarn test:unit
+```
