+++ src/components/DialogContainer.vue
@@ -13,7 +13,6 @@
     }
 </script>
 
-<!-- Add "scoped" attribute to limit CSS to this component only -->
 <style scoped>
   .modal-overlay {
     position: fixed;
